requirements.txt
asgiref==3.5.2
Django==4.0.5
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
sqlparse==0.4.2
